Module: Inventory

+++ src/main/java/com/sapo/store_management/dto/storage/StorageResponse.java
@@ -3,6 +3,7 @@ package com.sapo.store_management.dto.storage;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,8 +13,8 @@ public class StorageResponse {
     private String code;
     private String name;
     private String address;
-    private Long cash;
-    private Long balance;
+    private BigDecimal cash;
+    private BigDecimal balance;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
